상품 1. DTO Class inner 변환
