+++ src/main/java/com/ex/commercetestbackjpa/service/product/ProductService.java
@@ -25,18 +25,18 @@ public class ProductService {
     private final ProductPriceRepository productPriceRepository;
 
     @Transactional
-    public Long saveProduct(ProductRequestDto productRequestDto) {
+    public Long saveProduct(ProductDTO.Request productRequestDto) {
         Product product = productRequestDto.toEntity();
-        List<ProductDTRequestDto> productDTRequestDtoList = productRequestDto.getProductDtRequestDtos();
-        List<ProductPriceRequestDto> productPriceRequestDtoList = productRequestDto.getProductPriceRequestDtos();
+        List<ProductDtDTO.Request> productDTRequestDtoList = productRequestDto.getProductDtRequestDtoList();
+        List<ProductPriceDTO.Request> productPriceRequestDtoList = productRequestDto.getProductPriceRequestDtoList();
 
         Long productNo = productRepository.save(product).getProductNo();
 
-        for(ProductDTRequestDto productDTRequestDto : productDTRequestDtoList) {
+        for(ProductDtDTO.Request productDTRequestDto : productDTRequestDtoList) {
             productDtRepository.save(productDTRequestDto.toEntity(product));
         }
 
-        for(ProductPriceRequestDto productPriceRequestDto : productPriceRequestDtoList) {
+        for(ProductPriceDTO.Request productPriceRequestDto : productPriceRequestDtoList) {
             productPriceRepository.save(productPriceRequestDto.toEntity(product));
         }
 
@@ -44,9 +44,9 @@ public class ProductService {
     }
 
     @Transactional
-    public ProductResponseDto findProductByProductNo(Long productNo) {
+    public ProductDTO.Response findProductByProductNo(Long productNo) {
         Product product = productRepository.findById(productNo).orElseThrow(() -> new NoSuchElementException("조회된 상품이 없습니다."));
-        ProductResponseDto productResponseDto = new ProductResponseDto(product);
+        ProductDTO.Response productResponseDto = new ProductDTO.Response(product);
         productResponseDto.setProductDtResponseDtoList(this.addProductDtList(product));
         productResponseDto.setProductPriceResponseDtoList(this.addProductPriceList(product));
 
@@ -56,11 +56,11 @@ public class ProductService {
     @Transactional
     public Map<String, Object> findProductByProductName(String productName) {
         Map<String, Object> result = new HashMap<>();
-        List<ProductResponseDto> list = new ArrayList<>();
+        List<ProductDTO.Response> list = new ArrayList<>();
         List<Product> productList = productRepository.findByProductName(productName);
 
         for(Product product : productList) {
-            ProductResponseDto productResponseDto = new ProductResponseDto(product);
+            ProductDTO.Response productResponseDto = new ProductDTO.Response(product);
             productResponseDto.setProductDtResponseDtoList(this.addProductDtList(product));
             productResponseDto.setProductPriceResponseDtoList(this.addProductPriceList(product));
             list.add(productResponseDto);
@@ -73,11 +73,11 @@ public class ProductService {
     @Transactional
     public HashMap<String, Object> findProductAll() {
         HashMap<String, Object> result = new HashMap<>();
-        List<ProductResponseDto> list = new ArrayList<>();
+        List<ProductDTO.Response> list = new ArrayList<>();
         List<Product> productList = productRepository.findAll();
 
         for(Product product : productList) {
-            ProductResponseDto productResponseDto = new ProductResponseDto(product);
+            ProductDTO.Response productResponseDto = new ProductDTO.Response(product);
             productResponseDto.setProductDtResponseDtoList(this.addProductDtList(product));
             productResponseDto.setProductPriceResponseDtoList(this.addProductPriceList(product));
             list.add(productResponseDto);
@@ -88,32 +88,32 @@ public class ProductService {
         return result;
     }
 
-    private List<ProductDtResponseDto> addProductDtList(Product product) {
-        List<ProductDtResponseDto> productDtResponseDtoList = new ArrayList<>();
+    private List<ProductDtDTO.Response> addProductDtList(Product product) {
+        List<ProductDtDTO.Response> productDtResponseDtoList = new ArrayList<>();
         List<ProductDT> productDTList = product.getProductDtList();
 
         // 단품 add
         for (ProductDT productDT : productDTList) {
-            productDtResponseDtoList.add(new ProductDtResponseDto(productDT));
+            productDtResponseDtoList.add(new ProductDtDTO.Response(productDT));
         }
 
         return productDtResponseDtoList;
     }
 
-    private List<ProductPriceResponseDto> addProductPriceList(Product product) {
-        List<ProductPriceResponseDto> productPriceResponseDtoList = new ArrayList<>();
+    private List<ProductPriceDTO.Response> addProductPriceList(Product product) {
+        List<ProductPriceDTO.Response> productPriceResponseDtoList = new ArrayList<>();
         List<ProductPrice> productPriceList = product.getProductPriceList();
 
         // 가격 add
         for (ProductPrice productPrice : productPriceList) {
-            productPriceResponseDtoList.add(new ProductPriceResponseDto(productPrice));
+            productPriceResponseDtoList.add(new ProductPriceDTO.Response(productPrice));
         }
 
         return productPriceResponseDtoList;
     }
 
     @Transactional
-    public Long saveProductDt(ProductDTRequestDto productDTRequestDto) {
+    public Long saveProductDt(ProductDtDTO.Request productDTRequestDto) {
         Product product = productRepository.findById(productDTRequestDto.getProductNo()).get();
         ProductDT productDT = productDTRequestDto.toEntity(product);
 
@@ -122,7 +122,7 @@ public class ProductService {
 
     // 단품 색상 변경
     @Transactional
-    public Long updateProductDtColor(ProductDTRequestDto productDTRequestDto) {
+    public Long updateProductDtColor(ProductDtDTO.Request productDTRequestDto) {
         ProductDT productDt = productDtRepository.findById(productDTRequestDto.getProductDtNo()).orElseThrow(NoSuchElementException::new);
 
         productDt.updateColor(productDTRequestDto.getColorCode(), productDTRequestDto.getColorName());
@@ -131,7 +131,7 @@ public class ProductService {
     }
 
     @Transactional
-    public Long updateProductDtSize(ProductDTRequestDto productDTRequestDto) {
+    public Long updateProductDtSize(ProductDtDTO.Request productDTRequestDto) {
         ProductDT productDt = productDtRepository.findById(productDTRequestDto.getProductDtNo()).orElseThrow(NoSuchElementException::new);
 
         productDt.updateSize(productDTRequestDto.getSizeCode(), productDTRequestDto.getSizeName());
@@ -139,10 +139,10 @@ public class ProductService {
         return productDt.getProduct().getProductNo();
     }
 
-    public ProductDtResponseDto findProductDtByProductDtNo(Long productDtNo) {
+    public ProductDtDTO.Response findProductDtByProductDtNo(Long productDtNo) {
         ProductDT productDT = productDtRepository.findById(productDtNo).get();
 
-        return new ProductDtResponseDto(productDT);
+        return new ProductDtDTO.Response(productDT);
     }
